Refactored examples to use built version, and minify build output
