+++ package.json
@@ -40,7 +40,7 @@
     "eject-ts": "zx scripts/eject-typescript.mjs"
   },
   "dependencies": {
-    "vue-router": "^4.0.11"
+    "vue-router": "4.x"
   },
   "devDependencies": {
     "@babel/cli": "^7.12.1",
@@ -51,6 +51,7 @@
     "@storybook/vue3": "^6.3.12",
     "auto": "^10.3.0",
     "babel-loader": "^8.1.0",
+    "babel-preset-minify": "^0.5.1",
     "boxen": "^5.0.1",
     "concurrently": "^6.2.0",
     "dedent": "^0.7.0",
